+++ pyfilter/inference/sequential/nessmc2.py
@@ -1,7 +1,7 @@
 from .base import CombinedSequentialParticleAlgorithm
 from .ness import NESS
 from .smc2 import SMC2
-from ...kde import ConstantKernel, robust_var, ShrinkingKernel
+from ..kernels.kde import ConstantKernel, robust_var, ShrinkingKernel
 from ..utils import stacker
 from ...normalization import normalize
